+++ comics/comics/countyoursheep.py
@@ -6,18 +6,10 @@ class Meta(MetaBase):
     language = 'en'
     url = 'http://www.countyoursheep.com/'
     start_date = '2003-06-11'
+    end_date = '2011-12-07'
+    active = False
     rights = 'Adrian "Adis" Ramos'
 
 class Crawler(CrawlerBase):
-    history_capable_date = '2003-06-11'
-    schedule = None
-
-    # Without User-Agent set, the server returns 403 Forbidden
-    headers = {'User-Agent': 'Mozilla/4.0'}
-
     def crawl(self, pub_date):
-        page_url = 'http://countyoursheep.keenspot.com/d/%s.html' % (
-            pub_date.strftime('%Y%m%d'),)
-        page = self.parse_page(page_url)
-        url = page.src('img[src^="http://cdn.countyoursheep.keenspot.com/comics/"]')
-        return CrawlerImage(url)
+        pass # Comic no longer published