+++ docs/changes.rst
@@ -26,6 +26,7 @@ v1.1.6 (in development)
 - Update: ``boasas`` is no longer published.
 - Update: ``bunny`` is no longer published.
 - Update: ``carpediem`` is no longer published.
+- Update: ``countyoursheep`` is no longer published.
 - Update: ``dieselsweetiesweb`` to be more robust to missing elements in the
   feed.